==> ss_occ/01_prep_surv_covariates.R <==
# Create survey-level covariates 
# (see scripts in 'spatial/' for site-level covariates)

library(data.table)
library(dplyr)



### DID THE EFFORT COVARIATE IN '00_PREP_DETECTIONS' INSTEAD. WHAT ELSE DO I NEED? CALENDAR DATE? PROBABLY DO THAT THERE ALSO.
